+++ src/constants/settings.js
@@ -19,6 +19,8 @@ const settings = {
     root: "apiSettings",
     port: "apiSettings.port",
   },
+  singleInstance: "singleInstance",
+  disableHardwareMediaKeys: "disableHardwareMediaKeys",
   mpris: "mpris",
   enableCustomHotkeys: "enableCustomHotkeys",
   trayIcon: "trayIcon",